Develop

+++ src/main/java/integracionapp/psgtrading/dto/publisher/UserEvent.java
@@ -20,7 +20,7 @@ public class UserEvent {
         @AllArgsConstructor
         public class UserData {
             public String email;
-            public Integer VoteQuantity;
+            public Double VoteQuantity;
         }
     }
 }
